mcp-rag/bak-milvus-mcp-client/requirements.txt
mcp==1.5.0
dashscope==1.22.2
openai==1.76.0
markdown-it-py==3.0.0
markdownify==1.1.0
pdfplumber==0.11.6
sentence-transformers==2.4.0
python-dotenv==1.0.1
loguru==0.7.2 